+++ actiontext/test/test_helper.rb
@@ -22,6 +22,27 @@ if ActiveSupport::TestCase.respond_to?(:fixture_path=)
 end
 
 class ActiveSupport::TestCase
+  def assert_queries(num)
+    ActiveRecord::Base.connection.materialize_transactions
+    count = 0
+    queries = []
+
+    ActiveSupport::Notifications.subscribe("sql.active_record") do |_name, _start, _finish, _id, payload|
+      unless ["SCHEMA", "TRANSACTION"].include? payload[:name]
+        count += 1
+        queries << payload[:sql]
+      end
+    end
+
+    result = yield
+    assert_equal num, count, "#{count} instead of #{num} queries were executed. #{queries.inspect}"
+    result
+  end
+
+  def assert_no_queries(&block)
+    assert_queries(0, &block)
+  end
+
   private
     def create_file_blob(filename:, content_type:, metadata: nil)
       ActiveStorage::Blob.create_and_upload! io: file_fixture(filename).open, filename: filename, content_type: content_type, metadata: metadata